requirements.txt
absl-py==1.0.0
amqp==5.0.9
asgiref==3.4.1
astunparse==1.6.3
billiard==3.6.4.0
cachetools==4.2.4
celery==5.2.2
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cycler==0.11.0
Deprecated==1.2.13
Django==3.2.9
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-timezone-field==4.2.1
flatbuffers==2.0
fonttools==4.28.3
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kombu==5.2.2
libclang==12.0.0
Markdown==3.3.6
matplotlib==3.5.0
numpy==1.21.4

opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.24
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymemcache==3.5.0
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
setuptools-scm==6.3.2
six==1.16.0
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
tomli==1.2.2
typing_extensions==4.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
